+++ maro/simulator/scenarios/supply_chain/configs.py
@@ -97,7 +97,8 @@ test_world_config = {
                         "init_in_stock": 100,
                         "production_rate": 200,
                         "type": "production",
-                        "cost": 10
+                        "cost": 10,
+                        "price": 10
                     }
                 },
                 "storage": {
@@ -134,12 +135,16 @@ test_world_config = {
                         "init_in_stock": 100,
                         "production_rate": 200,
                         "type": "production",
-                        "cost": 10
+                        "cost": 10,
+                        "price": 100
                     },
                     # source material, do not need production rate
                     "sku3": {
                         "init_in_stock": 100,
-                        "type": "material"
+                        "production_rate": 200,
+                        "type": "material",
+                        "cost": 10,
+                        "price": 100
                     }
                 },
                 "storage": {
@@ -174,12 +179,15 @@ test_world_config = {
                 "skus": {
                     "sku1": {
                         "init_stock": 1000,
+                        "price": 100
                     },
                     "sku2": {
                         "init_stock": 1000,
+                        "price": 100
                     },
                     "sku3": {
                         "init_stock": 1000,
+                        "price": 100
                     },
                 },
                 "storage": {